Move gcp tests (#7318)
