+++ src/gcp/serviceusage.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import * as sinon from "sinon";
-import * as serviceUsage from "../../gcp/serviceusage";
-import * as poller from "../../operation-poller";
+import * as serviceUsage from "./serviceusage";
+import * as poller from "../operation-poller";
 
 describe("serviceusage", () => {
   let postStub: sinon.SinonStub;
